requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.9
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
motor==3.7.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pymongo==4.11.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
